refactor: extract hat tests
